# secp256k1: Avoid inversion in sig verify.